+++ jdaviz/configs/imviz/plugins/__init__.py
@@ -7,3 +7,4 @@ from .links_control import *  # noqa
 from .compass import *  # noqa
 from .aper_phot_simple import *  # noqa
 from .line_profile_xy import *  # noqa
+from .rotate_image import *  # noqa